feat: introduce lookup principle
Lookup is the principle that focuses on making sure we can find information after we document it.
As both our team and community grows, this is becoming more important to focus on.
Dendron gives us the tools to make this possible but it will be up to us to deliver on the promise.